+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### [13.36.1](https://kaos.sh/ek/13.36.1)
 
+- **`[spinner]`** Description now limited by size of the line
 - **`[terminal/tty]`** Fixed bug with closing tty file descriptor
 
 ### [13.36.0](https://kaos.sh/ek/13.36.0)